requirements.txt
alembic==1.7.5
asgiref==3.3.4
autopep8==1.5.7
bcrypt==3.2.0
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.65.2
greenlet==1.1.0
h11==0.12.0
httptools==0.2.0
idna==3.2
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2==2.9.1
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
python-dotenv==0.18.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.19
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
uvloop==0.15.2
watchgod==0.7
websockets==9.1
